第4次提交

+++ exercises/23_conversions/using_as.rs
@@ -3,6 +3,7 @@
 // Type casting in Rust is done via the usage of the `as` operator. Please note
 // that the `as` operator is not only used when type casting. It also helps with
 // renaming imports.
+// Rust中的类型转换是通过使用“as”运算符完成的。请注意，“as”运算符不仅在类型转换时使用。它还有助于重命名导入。
 //
 // The goal is to make sure that the division does not fail to compile and
 // returns the proper type.
@@ -10,11 +11,9 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
